+++ db/migrate/20160805170059_remove_data_type_from_answer_templates.rb
@@ -1,0 +1,5 @@
+class RemoveDataTypeFromAnswerTemplates < ActiveRecord::Migration[5.0]
+  def change
+    remove_column :answer_templates, :data_type, :string
+  end
+end