+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - Removed unused user contact information
 - Cleaned up unused survey JavaScript
 - Removed deprecated display_type column from questions
+- Removed deprecated data_type column from answer_templates
 
 ### Bug Fix
 - Fixed a bug that allowed users to be assigned the same survey multiple times